Adding method to auto-generate interaction model file
Added method to generate the interaction model, and an Explicit test that will run the method and write the results to a file so it can be added to the skills Interaction Model in the console.  This method has no bearing on how the skill works, but makes it easier to generate the supporting files.